Tests demonstrating a failure to call methods on P5 objects
Chaining method calls on objects returned by $i.call works, but saving
the object into a variable and calling methods on that fails with
add_fallback trying to add a fallback for "call". Seels like $!perl5
is a Perl5Object instead of a PerlInterpreter in this case.

Can be fixed by making all Perl5Object attributes public and removing
the BUILD method.