Fix segfaulting on exceptions in P5 functions.
If a P5 function or method doesn't return any values or gets interrupted
by an exception, we called av_extend with a negative count. Strangely
enought this doesn't bother Perl 5, except when doing this during a
presentation.

The only known test case is starting a Catalyst application when the
port is already taken. Failed to provoke it under any other
circumstances, hence no tests for this. The failure condition occurs in
many tests but never leads to an explosion.