Fix passing empty strings and bufs to Perl 5.
A 0 len parameter passed to newSVpv makes perl5 use strlen() to
determine len. We don't need this magic anymore, since we determine the
length of the buffer in Perl 6 code already for binary safety.

Works around an apparent bug in Rakudo. The pointer to the supposedly
empty string points to random garbage now.