Fix method calls on P5 exception objects
When we started supporting stringification of Perl 5 objects,
S01-perl-5-integration/exception_handling.t started failing.

The reason is that the ERRSV pointer never changes. Instead, the SV it
points to gets overwritten with the exception object. As a result, we
cannot just wrap the ERRSV pointer in a Perl5Object, since this only
works until Perl 5 resets $@. From that point on all method calls will
fail.

Fixed by creating a copy of the SV before converting it to a Perl 6
scalar.