Hopefully fix the rare panic: MUTEX_LOCK in t/callables.t
hoelzro++ finally discovered, what's going on here. His diagnosis:

The issue seems to be that the Inline::Perl5 DESTROY submethod tears down
the Perl 5 interpreter and then the DESTROY submethod on a Perl5Callable
is invoked. The global data structures that perl uses (such as its op
mutex) have been taken down and are not in a usable state, so
pthread_mutex_lock sees the misconfigured mutex and freaks out, returning
EINVAL (which is the 22 we see)

Fix by having p5_destruct_perl set a "destroyed" flag that p5_sv_refcnt_dec
checks. This is kind of a hot path, but doing this in C, I guess it will
still not have measurable performance impact. At least until we get better
performance out of NativeCall.

Fixes GH #34