Detect perls not configured with a shared library
Inline::Perl5 can only work if perl was configured with useshrplib; if
a perl without this configuration option is used, the user gets a host
of compiler errors that they may or may not understand.  Bailing out
early in the build process tells the user exactly what is going on,
and what they need to do to fix it.