Hide Perl5Parent from v6-inline users
When using v6-inline to extend a Perl 5 package in Perl 6, one had to explicitly
also does Perl5Parent['Foo::Bar'];
and pass the Inline::Perl5 object to the constructor.

Both are now gone. Since we already assume that the user wants to create a
class of the same name as the Perl 5 package, we may as well assume that this
class should extend the Perl 5 package. Since they are both in the same file,
this  should be a fair assumption.

In the place where we create the Perl 6 class, we also know exactly, which
Perl 5 interpreter the Perl 5 class is from, so we pass it as Role parameter
instead of requiring the user to pass it to the constructor.