New v6::extend helper for easier extending of P5 classes with v6-inline
use v6-inline; can be used to write part of a Perl 5 class in Perl 6. When
Perl 5 code should be able to create objects of this class, a bit of a
dance is necessary. A Perl 5 constructor creates the Perl 5 object and
calls into Perl 6 to create the corresponding Perl 6 object. Those two
are then hooked together by some reblessing.

v6::extend is the generic version of this dance, so that the typical
Perl 5 construtor call can be like:
sub create {
    my ($class, %args) = @_;
    return v6::extend($class, $class->new(%args), [], \%args);
}