Better fake inheritance from P5 classes by P6 classes
We now always have a representation of the P6 subclass in Perl 5.
Instead of using trickery to call P5 methods with a wrapped P6 object as
$self, we re-bless the P5 object into the Perl 5 representation of the
Perl 6 subclass and always pass the original P5 object around.

This way it can no longer happen that a Perl 5 function gets the wrapped
P6 object when it's expecting something that looks like the Perl 5 base
object.