Fix one source of "Attempt to free unreferenced scalar during global destruction"
This was just left over code from before the inheritance re-design.
No need to create a new reference to bless the object. We can just bless
the reference we were passed.

The "unreferenced scalar" error came, because we used newRV_noinc to
create the reference and then never increased or decreased the refcount
afterwards. Without an SvREFCNT_dec, perl never gets a chance to free
the SV and rightfully complains about that when it finds it during
shutdown.