Fix calling imported functions that don't exist in the source package
When a loaded module exported a symbol that doesn't exist in the module
itself because it was for example generated at runtime or pulled from
some other module, the call would fail because we always looked in the
loaded module directly. Just call the imported function instead.

This fixes Dancer2's DSL.

Thanks to tadzik++ for pointing this out!