Avoid slurpies for less overhead when calling P5 methods
By passing on the Capture instead of packing and unpacking slurpies,
we can win another 5 % in csv-ip5xs.
Previously an Empty returned by a P5 function without return values was
filtered by the interpolation of the slurpies. Now we have to explicitly
return a Nil instead (which is more correct anyway), turn it into a NULL
and filter that in C code.

Version 0.17