Replace use of Perl5Parent for v6::inline by Perl5Extension
This first part adds the v6::code package for using a different way of
extending a Perl 5 package with Perl 6 code. Instead of abusing inheritance,
we create Perl 5 wrappers for Perl 6 methods and subs.

On the Perl 6 side, Perl 6 extended objects are indistinguishable from plain
Perl 5 objects. On the Perl 5 side, the only way to notice anything out of
the norm is by looking for the attached magic.

Fixes memory leaks caused by the Perl 5/6 halves of extended objects to
point to their other half, keeping both alive. Now the Perl 5 object is the
authoritative source.