Speed up repeated calls to methods on Perl 5 objects
Method fallbacks have considerable overhead. But once we know that a
Perl 5 object may have some method, we can assume that its gonna be
called again in the future. So we install a wrapper method with the name
we now know to avoid the overhead of fallbacks in future calls.