Fix method calls on P5 objects swallowing named args
Unfortunately named arguments play only a very minor role in multi dispatch,
so the presence of a candidate taking a Capture won't help and other candidates
will still be called. So we still have to check %_ and manually create the
Capture.

This will cost some performance but fortunately still retains some of the
benefit of creating multis at runtime.