Ensure forwards compatibility
We use nqp ops for handling the multi dispatch cache. Those are not a stable
API and can go away any time. To prepare for that event, we compile the
find_best_dispatchee method in an EVAL and fall back to unoptimized code in
case compilation fails.
Note that this will just keep things running, or rather crawling. Performance
will be pretty terrible.