Fix runaway refcount on method calls on objects of P5 wrapper classes
This really just reverts part of "Properly represent P6 subclasses of P5
packages in P5". I don't know why I did that change in the first place. It
looks like an artifact of experimentation that I overlooked when doing the
commit.

The invocant is treated specially (not mortalized) by p5_call_gv. Thus we
must not increase its refcount before doing the call.