Fix leaking objects of wrapped P5 class
Objects of wrapped Perl 5 classes (Inline::Perl5::ClassHOW) consist of
two parts: the Perl 5 object of the Perl 5 (base) class and a Perl 6
object referencing it and potentially holding additional attributes and
methods. At any point in time the combined object may be referenced in
Perl 5, in Perl 6 or in both. This means that neither half can dictate
the life time of the other. Both objects may only go away when neither
of them is in use anymore. Since this is a circular reference with no
way to detect it, we leaked all of these objects.

Unfortunately there is no way the garbage collectors of the Perl 6 and
Perl 5 implementations can cooperate on this. Luckily in Perl 5
destructors can resurrect an object by creating new references to it.
The same is actually true in Perl 6, but there the destructor will be
called only a single time over the full life time of the object. So the
object could only be resurrected once.

We now use the destructor to get information about whether the object is
still referenced by any Perl 5 code or data. When the object gets
destroyed and resurrected, we clear its reference to the Perl 6 half, so
that may be garbage collected later on. When instead the object gets
passed to Perl 5 code again, we reinstate the reference to the Perl 6
half, keeping it alive.

While this elaborate scheme fixes the memory leak, it costs about 28 %
performance in the standard benchmark.