Speed up common case of P5 half of object needing reset
When calling a method on a wrapped object (Inline::Perl5::ClassHOW), we
know that it is referencing its Perl 6 half. Before decreasing the
refcount at the end of the call, we can check if it would reach 0. In
that case we can avoid the call to the Perl 5 destructor and directly
short circuit to what p5_destroy_p5_object would do. This avoids some of
the performance penalty of the whole destructor/resurrection dance.