Fix spurious test failures on 32 bit machines
On 32 bit machines pointers are only half as large as on 64 bit machines which
means that the GC threshold will be reached later. So add more pressure on the
GC by allocating lots of memory that will end up in the old generation. This
makes the tests pass much faster, too.