Fix occasional endless DESTROY loop at end of use_nested.t
Getting a reference to a function by full name will always give a reference,
even if there was no function by that name before. The defined() function is
actually special cased, so defined(&func) will not call the function, but
instead can be used to check for the existence of the function. Use this to
determine whether the class already has a destructor, avoiding accidental
call of our implanted destructor from itself.