Fix leaking hashes passed to Perl 5 code
We need to mortalize our reference to the hash as the Perl6::Hash constructor
will create its own. Otherwise our reference would never be freed, keeping the
hash and the Raku hash alive. In addition we leaked the scalar wrapping the
"Perl6::Hash" string every time we wrapped a hash.