Fix leaking handles passed to Perl 5 code
We need to mortalize our reference to the handle as the Perl6::Handle constructor
will create its own. Otherwise our reference would never be freed, keeping the
handle and the Raku handle alive. In addition we leaked the scalar wrapping the
"Perl6::Handle" string every time we wrapped a handle.