Fix loading Math::BigFloat:from<Perl5>
Math::BigFloat failed to load with Not a HASH reference at
Math::BigFloat::is_zero because the code looking for variables in the package
caused evaluation of the variable in boolean context and since Math::BigFloat
tied one such variable this caused the boolification overload to be called.
The overloaded sub did not cope too well with the string it got called on.
Fixes GH #158