Fix thread interpreters missing modules after precompilation
Loading of modules happens during precompilation. But each module is
precompiled in a separate process, so the Inline::Perl5 objects that hold the
list of required modules is a different one each time. So while we actually
loaded all required modules, we didn't know the full list of modules we loaded
and couldn't use it to prime new thread interpreters.