Don't return Scalar containers from find_method
Apparently when trying to get things to work I made find_method is raw at some
point and didn't check if that was actually necessary. Either that or at some
point it actually was necessary (for whatever reason). Anyway it's not needed
now and there's no good reason to return Scalar containers. Quite contrary,
that can only hurt if the caller doesn't decont.