Fix installation on new-disp
The trick with 3 different modules implementing a find_best_dispatchee sub
worked during compilation of Inline::Perl5::ClassHOW. Unfortunately it doesn't
work when one tries to install the distribution, because there the error to
compile the Full version causes installation to abort.

Instead of using modules to contain any compilation errors, we now run a
separate compilation process manually and if that succeeds, EVAL the code in
place.

It would be more elegant to just catch any exceptions thrown by an EVAL, but
that leads to other errors as EVAL during BEGIN time will modify the state
of the outer compiler and the exception leave that state inconsistent.