Fix Perl exception objects getting swallowed in callbacks
When Raku calls into Perl which calls into Raku which finally calls into Perl
and the latter throws an exception object, along the way back to the outer
Raku code this object got turned into an empty string which was then no longer
recognized as an exception.
Wrap such objects in Inline::Perl5::Exception objects which are derived from
Exception and delegate stringification.