Fix rare "Cannot convert string to number" error in ObjectKeeper
While the ObjectKeeper is thread safe (by virtue of having one per Perl
interpreter) it is not actually reentrant. This matters when during a call
to ObjectKeeper.keep a garbage collection run takes place and this run causes
a Raku object to be freed that holds the last reference to a Perl object which
reference another Raku object. The latter would then be removed from the
ObjectKeeper. If this happens between reading $!last_free and updating
$!last_free with the next location read from the free-chain we end up writing
the object to keep into a position that has just become part of the free-chain.

Fix by making sure we only read $!last_free once, so we get a consistent
result. The downside of this simple fix is that instead of crashing we
actually lose a slot in the objects array. This slot is the one that would have
been pointed to by $!last_free, but is now missing from the free-chain.
However, since this is incredibly rare, it's not clear if a more comprehensive
solution is really necessary.