Allow for existence of POPULATE method as well
As Fritz Zaucker showed in

  https://gist.github.com/zaucker/5800adb925e3836259387a00f5a717bc

Inline::Perl5 would segfault on a simple method call, which was
traced back to the introduction of POPULATE in

  https://github.com/rakudo/rakudo/commit/ea1831c79f

This commit adds POPULATE to the list of expected methods in Mu,
and conditionally puts it in the cache if it exists (which also
applies to the other methods mentioned in that list, which shouldn't
be an issue.

Tested both on pre-POPULATE rakudo, as well as post 2025.03 main.